backend/pyproject.toml
[tool.poetry]
name = "replant"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.1"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"
colorlog = "^6.8.0"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
whitenoise = "^6.6.0"
django-debug-toolbar = "^4.2.0"
pycountry = "^23.12.11"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
django-filter = "^23.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.7.0"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}
django-types = "^0.19.1"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
pytest-xdist = "^3.5.0"
model-bakery = "^1.17.0"
time-machine = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "replant.settings"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:The django.utils.baseconv module is deprecated",
]

[tool.black]
exclude = "(.venv|migrations)"
target_version = ["py310"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip = [".venv", "migrations"]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = "migrations"
ignore_missing_imports = true
check_untyped_defs = true
show_column_numbers = true
warn_unused_ignores = true
strict_equality = true
no_implicit_optional = true
follow_imports = "silent"
mypy_path = "backend"

[tool.django-stubs]
django_settings_module = "replant.settings"
